+++ uplt/engine/PlotlyEngine5.py
@@ -1,20 +1,17 @@
 import importlib.util
-from uplot.interface import IPlotEngine, IFigure
-from uplot.default import DEFAULT
+from uplt.interface import IPlotEngine, IFigure
+from uplt.default import DEFAULT
 
 
 class PlotlyEngine5(IPlotEngine):
-    # engine specific default parameters
-    FILE_RESOLUTION_SCALE = 2
-    LINE_WIDTH = 2.5
 
     @property
     def name(self) -> str:
-        return 'plotly5'
+        return 'plotly'
 
     @classmethod
     def is_available(cls) -> bool:
-        return importlib.util.find_spec("plotly") is not None
+        return importlib.util.find_spec('plotly') is not None
 
     @property
     def go(self):
@@ -34,15 +31,15 @@ class PlotlyEngine5(IPlotEngine):
 
         # load style
         if DEFAULT.style.lower() == 'bmh':
-            from uplot.engine.style.plotly import bmh
+            from uplt.engine.style.plotly import bmh
             self._layout_style = bmh
         else:
             raise NotImplementedError(f'style not supported for plotly: {DEFAULT.style}')
 
-    def figure(self, width: int, aspect_ratio: float) -> IFigure:
-        from uplot.engine.PlotlyFigure5 import PlotlyFigure5
 
-        fig = PlotlyFigure5(self)
+    def figure(self, width: int, aspect_ratio: float) -> IFigure:
+        from uplt.engine.PlotlyFigure5 import PlotlyFigure5
+        fig = PlotlyFigure5(engine=self) # type: ignore
 
         # adjust style layout
         fig.internal.update_layout(template=self._layout_style,