+++ uplt/interface.py
@@ -5,8 +5,8 @@ from typing import Any, Protocol, runtime_checkable
 from abc import abstractmethod as abstract
 from numpy.typing import ArrayLike
 
-from uplot.utype import LineStyle, MarkerStyle, AspectMode, AxisScale, Colormap
-from uplot.utool import Interpolator
+from uplt.utype import LineStyle, MarkerStyle, AspectMode, AxisScale, Colormap
+from uplt.utool import Interpolator
 
 
 @runtime_checkable
@@ -566,7 +566,7 @@ class IFigure(Protocol):
         IFigure
             The figure object representing the plot.
         """
-        
+
     @abstract
     def current_color(self) -> str:
         """
@@ -709,4 +709,3 @@ class IPlotEngine(Protocol):
         IFigure
             A new figure instance.
         """
-